# Add comprehensive PDF validation system with enhanced error handling