--- a/api/validators/__init__.py
+++ b/api/validators/__init__.py
@@ -1,0 +1,12 @@
+"""
+PDF Validation Module
+
+This module provides PDF validation functionality to check if PDFs
+can be processed before attempting extraction.
+"""
+
+from .pdf_validator import PDFValidator
+from .error_codes import ErrorCode
+from .validation_result import ValidationResult, PDFType
+
+__all__ = ['PDFValidator', 'ErrorCode', 'ValidationResult', 'PDFType']