… docs/PDF_ENHANCEMENT_PLAN.md
… tests/test_pdf_validator.py